Added combat arts page and fixed all links errors
All links related to js files were broken and now they are fixed.

The combat arts plugin was updated to version 3.0, so the js file that was handling the download of said plugin was changed to account for new files.

A new html page was created to explain the combat arts plugin. In the next update it is expected that it will be edited to add screenshots to help explain it.